Telecom: Add support for call timer reset on CDMA MO call
In android original design, the duration of CDMA MO call is
started from the dial command sent for some CDMA carriers, so
it is not the real duration of the active time.

In this patch, a new message is registered to listen the event
of the call accepted, and then reset the duration when the
event happens.

Change-Id: Icc447012030ae243f200ec2c83b7d5210af9b31c
CRs-Fixed: 2207419
Signed-off-by: VenkatVishalV <venkatvishal124@gmail.com>
Signed-off-by: sayan7848 <sayan7848@gmail.com>