Fix the behavior of keyguard bouncer in a corner case
With Android Automotive, there's a cancel button on the bouncer
that allows return to the UserSwitcher and selection to another user.
This new user may have different security mode which requires bouncer
container to be resized. As a result, the bouncer view will be destroyed.
So keyguard bouncer should reset its security status when keyguard view
is resumed.

Test: 1. set screen lock with pin, password or pattern
Test:    in Settings ->Security & location
Test: 2. use command "adb shell input keyevent 26 " to lock/unlcok device
Test: 3. input wrong pin or password for 5 times
Test: 4. select cancel button when wait for 30 seconds
Test: 5. enter lock screen and input wrong password again

bug: 119068682
Change-Id: I092f1f39def87a679847872507653319ebd1779d
Signed-off-by: sayan7848 <sayan7848@gmail.com>